Revisão das pastas e includes
Corrige a organização dos diretórios e as páginas de include para
adicionar os cabeçalhos e rodapés com o PHP. Adiciona o link no botão
sair, cria uma pasta logo para organização. Muda a apresentação do
perfil para o nome do usuário.
