--- assets/userImages/README.md
@@ -1,1 +1,0 @@
-# Pasta de fotos de perfil dos usuários do site
